+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 
 ## 1.1.0
 - Has image loading widget for multiple scenarios like svg loading, cached network.
+
+## 1.2.0
+- Docs about Container_ widget
+- Example has a new screen with all features of Container_ widget